+++ src/decorators/componentRedirectRoute.decorator.ts
@@ -1,6 +1,7 @@
 import {Route} from '@angular/router';
+import {isBlank, isPresent} from '@js/common';
+
 import {RouteDecoratedComponent} from './componentRoute.decorator';
-import {isBlank, isPresent} from '../utils/lang';
 
 /**
  * Extended type with redirect route definition