+++ src/services/cookies/cookies.service.ts
@@ -1,8 +1,8 @@
 import {PLATFORM_ID, Inject, Optional, Injectable} from '@angular/core';
 import {isPlatformBrowser} from '@angular/common';
+import {isBlank} from '@js/common';
 
 import {SERVER_COOKIE_HEADER} from '../../types/tokens';
-import {isBlank} from '../../utils/lang';
 
 /**
  * Class that is used as wrapper for working with cookies