+++ src/types/dictionaries.ts
@@ -1,2 +1,0 @@
-export type Dictionary = {[index: string]: any};
-export type StringDictionary = {[index: string]: string};