+++ src/utils/utils.ts
@@ -1,32 +1,10 @@
 import {NgForm, FormGroup} from '@angular/forms';
 import {NgModuleRef} from '@angular/core';
 import {Route} from '@angular/router';
-import Encoder from './utils.encoder';
+
 import RouterHelper from './utils.routerHelper';
 import Common from './utils.common';
 import Forms from './utils.forms';
-
-/**
- * Used for encoding/decoding strings
- */
-export interface IEncoder
-{
-    /**
-     * Decode html encoded string
-     * @param  {string} input String to be decoded
-     * @return string
-     */
-    htmlDecode(input: string): string;
-
-
-    /**
-     * Encodes html plain string
-     * @param  {string} input String to be encoded
-     * @param  {boolean} dbl Indication that double encoding should be performed
-     * @returns string
-     */
-    htmlEncode(input:string, dbl?: boolean): string;
-}
 
 /**
  * Helper methods for angular2 router
@@ -46,42 +24,6 @@ export interface IRouterHelper
  */
 export interface ICommon
 {
-    /**
-     * Reverse current string and returns new reverse string
-     * @param  {string} str String to be reversed
-     * @returns string Reverse string
-     */
-    reverseString(str: string): string;
-    
-    /**
-     * Extends one object with additional properties from other objects, supports deep extend
-     * @param  {boolean|Object} deepOrObject Object to be extended or indication that deep copy should be performed
-     * @param  {Object[]} objectN Objects that will be used for extending, if deep is used first here is target object
-     * @returns Object Extended object with properties from other objects
-     */
-    extend(deepOrObject: boolean | Object, ...objectN: Object[]): Object;
-    
-    /**
-     * Merges properties of two separate object into new third one
-     * @param  {{[key: string]: any}} source1 First source object 
-     * @param  {{[key: string]: any}} source2 Second source object
-     * @returns Object Object containing properties from source1 and source2 objects
-     */
-    merge(source1: {[key: string]: any}, source2: {[key: string]: any}): Object;
-    
-    /**
-     * Generates random string consisting from lowercase letters
-     * @param  {number} length Length of generated string
-     * @returns number Generated string
-     */
-    generateId(length: number): string;
-
-    /**
-     * Converts string in that way that first letter will be lowerCase
-     * @param  {string} text Text to be converted
-     */
-    firstToLowerCase(text: string): string;
-
     /**
      * Runs callback function when angular module is bootstrapped and stable
      * @param {Promise<NgModuleRef<{}>>} moduleRefPromise Promise for module that was bootstrapped
@@ -135,14 +77,6 @@ export interface IForms
  */
 class Utils
 {
-    /**
-     * Encoder of strings
-     */
-    static get encoder(): IEncoder
-    {
-        return new Encoder();
-    }
-    
     /**
      * Angular 2 router helper
      */